+++ CHANGELOG.md
@@ -3,7 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-Nothing yet
+### Changed
+- [#52](https://github.com/Kashoo/synctos/issues/52): Upgrade development dependencies
 
 ## [1.3.0] - 2016-11-23
 ### Added