+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [1.3.1] - 2016-11-24
 ### Changed
 - [#52](https://github.com/Kashoo/synctos/issues/52): Upgrade development dependencies
 
@@ -33,7 +33,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - [#36](https://github.com/Kashoo/synctos/issues/36): Does not return a non-zero exit status when sync function generation fails
 
-[Unreleased]: https://github.com/Kashoo/synctos/compare/v1.3.0...HEAD
+[Unreleased]: https://github.com/Kashoo/synctos/compare/v1.3.1...HEAD
+[1.3.1]: https://github.com/Kashoo/synctos/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/Kashoo/synctos/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/Kashoo/synctos/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/Kashoo/synctos/compare/v1.0.0...v1.1.0